270121-10:00
Added code which enables the timer to measure accurately in seconds regardless of CPU cycle rate.

Added code which adjusts the thrust vector of the satellite based on its angle (needs implementation with rotational mechanics still).

Added code to the input demo in PESO_main.cpp which enabels the user to specify the roll, yaw and pitch of the satellite. Position has been redesigned to represent distance from Earth, so as to give the user a reference point as to where they would like to position their satellite prior to simulation runtime.

Added code which provides an authentic orthographic 3D experience - prior to this there was a bug involving the YZ viewport - the y-axis was assigned the vertical value and the z-axis the horizontal, where respectively they should represent the converse due to their viewing point.