240221-17:00
Added SDL_Delays to filesys debug messages to enable the user to see what's happening in the background when closing down PESO's executable file.

Executable wouldn't run without crashing because the 'res' folder was not copied into the build/Debug directory - causing the application font to be SDL_NULL and cause an early exit.